+++ src/com/farmacia/join_entidades/JoinListarNotaPedidosCabecera.java
@@ -1,6 +1,7 @@
 
 package com.farmacia.join_entidades;
 
+import java.math.BigDecimal;
 import java.sql.Date;
 import java.sql.Time;
 
@@ -19,14 +20,14 @@ public class JoinListarNotaPedidosCabecera {
     private String Plazo;
     private String forma_pago;
     private String fecha_creacion;
-    private Double iva;
-    private Double descuento;
-    private Double total;
+    private BigDecimal iva;
+    private BigDecimal descuento;
+    private BigDecimal total;
 
     public JoinListarNotaPedidosCabecera() {
     }
 
-    public JoinListarNotaPedidosCabecera(Long id_cabecera_nota_pedidos, Long id_proveedor, String entidad, String representante, String telefono, Long id_proveedor_clase, String clase, String direccion, String cedula_ruc, String correo, String Plazo, String forma_pago, String fecha_creacion, Double iva, Double descuento, Double total) {
+    public JoinListarNotaPedidosCabecera(Long id_cabecera_nota_pedidos, Long id_proveedor, String entidad, String representante, String telefono, Long id_proveedor_clase, String clase, String direccion, String cedula_ruc, String correo, String Plazo, String forma_pago, String fecha_creacion, BigDecimal iva, BigDecimal descuento, BigDecimal total) {
         this.id_cabecera_nota_pedidos = id_cabecera_nota_pedidos;
         this.id_proveedor = id_proveedor;
         this.entidad = entidad;
@@ -149,29 +150,28 @@ public class JoinListarNotaPedidosCabecera {
         this.fecha_creacion = fecha_creacion;
     }
 
-    public Double getIva() {
+    public BigDecimal getIva() {
         return iva;
     }
 
-    public void setIva(Double iva) {
+    public void setIva(BigDecimal iva) {
         this.iva = iva;
     }
 
-    public Double getDescuento() {
+    public BigDecimal getDescuento() {
         return descuento;
     }
 
-    public void setDescuento(Double descuento) {
+    public void setDescuento(BigDecimal descuento) {
         this.descuento = descuento;
     }
 
-    public Double getTotal() {
+    public BigDecimal getTotal() {
         return total;
     }
 
-    public void setTotal(Double total) {
+    public void setTotal(BigDecimal total) {
         this.total = total;
     }
 
-    
 }
